attackSpeed added
now you cant spam the attack, you need to wait depends on the attackspeed
fixed arrow shoot position

+++ King_of_the_Hill/King_of_the_Hill/Model/GameItems/Axe.cs
@@ -3,7 +3,7 @@
     using System.Drawing;
     public class Axe : Weapon
     {
-        public Axe(double weaponDamage, string weaponName, double attackSpeed, double PosX, double PosY) : base(weaponDamage, weaponName, attackSpeed, PosX, PosY)
+        public Axe(double weaponDamage, string weaponName, int attackSpeed, double PosX, double PosY) : base(weaponDamage, weaponName, attackSpeed, PosX, PosY)
         {
             
         }
