+++ King_of_the_Hill/King_of_the_Hill/Model/GameItems/Sword.cs
@@ -2,7 +2,7 @@
 {
     public class Sword : Weapon
     {
-        public Sword(double weaponDamage, string weaponName, double attackSpeed, double PosX, double PosY) : base(weaponDamage, weaponName, attackSpeed, PosX, PosY)
+        public Sword(double weaponDamage, string weaponName, int attackSpeed, double PosX, double PosY) : base(weaponDamage, weaponName, attackSpeed, PosX, PosY)
         {
 
         }
