+++ King_of_the_Hill/King_of_the_Hill/Model/GameItems/Weapon.cs
@@ -5,9 +5,9 @@
     public class Weapon : GameItem
     {
         public double WeaponDamage { get; set; }
-        public double AttackSpeed { get; set; }
+        public int AttackSpeed { get; set; }
 
-        public Weapon(double weaponDamage, string weaponName, double attackSpeed, double PosX, double PosY) : base(weaponName, PosX, PosY)
+        public Weapon(double weaponDamage, string weaponName, int attackSpeed, double PosX, double PosY) : base(weaponName, PosX, PosY)
         {
             WeaponDamage = weaponDamage;
             Name = weaponName;
